refactor(agent): extract agent orchestration into dedicated crate
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
