+++ crates/agent/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "agent"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+publish.workspace = true
+
+[dependencies]
+acrawl-core = { path = "../core" }
+base64 = "0.22"
+browser = { path = "../browser" }
+csv = "1"
+regex = "1"
+runtime = { path = "../runtime" }
+serde_json = "1"
+tokio = { version = "1", features = ["sync", "time", "fs"] }
+
+[dev-dependencies]
+async-trait = "0.1"
+tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
+
+[lints]
+workspace = true
