+++ crates/agent/src/agent.rs
@@ -1,0 +1,4 @@
+#[path = "implementation/mod.rs"]
+mod implementation;
+
+pub use implementation::*;
