+++ crates/agent/src/manager.rs
