+++ crates/agent/src/output.rs
