+++ crates/agent/src/registry.rs
