+++ crates/agent/src/shared_client.rs
