+++ crates/agent/src/state.rs
