+++ crates/agent/src/tools/click.rs
