+++ crates/agent/src/tools/execute_js.rs
