+++ crates/agent/src/tools/fork.rs
