+++ crates/agent/src/tools/hover.rs
