+++ crates/agent/src/tools/list_resources.rs
