+++ crates/agent/src/tools/navigate.rs
