+++ crates/agent/src/tools/read_content.rs
