+++ crates/agent/src/tools/screenshot.rs
