+++ crates/agent/src/tools/scroll.rs
