+++ crates/agent/src/tools/select_option.rs
