+++ crates/agent/src/tools/switch_tab.rs
