+++ crates/cli/src/app/mod.rs
@@ -23,7 +23,7 @@ use crate::output_sink::{ChannelSink, OutputSink, StdoutSink};
 use crate::session_mgr::{create_managed_session_handle, SessionHandle};
 use crate::tui::ReplTuiEvent;
 use commands::{slash_command_specs, SlashCommand};
-use crawler::mvp_tool_specs;
+use agent::mvp_tool_specs;
 use runtime::{
     CompactionConfig, ControlState, ConversationRuntime, RuntimeError, Session, TokenUsage,
 };
