+++ crates/crawler/Cargo.toml
@@ -6,20 +6,16 @@ license.workspace = true
 publish.workspace = true
 
 [dependencies]
-acrawl-core = { path = "../core" }
+agent = { path = "../agent" }
 browser = { path = "../browser" }
-runtime = { path = "../runtime" }
-async-trait = "0.1"
-rand = "0.8"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-socket2 = "0.5"
-tokio = { version = "1", features = ["fs", "io-util", "macros", "net", "process", "rt", "sync", "time"] }
-tokio-tungstenite = "0.26"
+acrawl-core = { path = "../core" }
+
+[dev-dependencies]
 futures-util = "0.3"
-csv = "1"
-base64 = "0.22"
-regex = "1"
+runtime = { path = "../runtime" }
+serde_json = "1"
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net"] }
+tokio-tungstenite = "0.26"
 
 [lints]
 workspace = true
